Split the original `HubController` in a controller folder `Hub`
