+++ app/Http/Controllers/Hub/HubController.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Hub;
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\View\View;
+use App\Http\Controllers\Controller;
 
 class HubController extends Controller
 {
